Suppress cppcheck warnings about comparing pointers
`kernel_start`, `kernel_end`, `isr_source` and `isr_end` are set up in
the linker file, so when parsing the c++ it looks like we are comparing
totally unrelated pointers. However, we do need to get the size of the
isr section and the whole kernel, as well as iterate through each of the
isr handlers in a couple places.