Avoid moving the exceptions table to 0x0.
Set the Raspberry Pi's Vector Base Address Register to where the table
is so we don't have to copy it to 0x0 at runtime (which meant doing a
little dance to avoid overwriting the device tree).
Also, since we couldn't load code directly to 0x0, the PMM couldn't know
where it was from the linker file so we had to hardcode it. This way
there is one less hardcoded memory location in the PMM!