Fix exception label misalignment.
The exceptions themselves were aligned, but the address that was loaded
into the control register was set in the linker file and not aligned.
The label is now moved to be in the same file as the code it's
addressing now that we no longer need to relocate it at runtime.