Use a linked-list for the PMM instead of a bitmap
Currently, items are never split or joined, so you only have as many
allocations as ranges of memory given by the bootloader. Also, headers
are stored seperate from the memory, but there is no code to store more
of them.
Memory is still handed out with a granularity of 4_KiB (not changing).