Actually use the PMM linked list to track memory
Now it gets split into correct-sized chunks (4_KiB granularity) when
being handed out, and merged back together when freed.

I'm using a linked list of arrays of linked-list items. So you have a
linked list where each item takes up 4_KiB memory, and has just a next
pointer and a bunch of structs that form a seperate linked list. This
allows the PMM to only need to self-request memory in 4_KiB chunks,
while still having many dynamic structs.
The second linked-list is ordered by address, with no overlaps (but
possibly gaps). Each entry is either in-use or not. An item not part of
the chain as the header_in_use bool unset so it can be added safely
without having to allocate anything new.
To allocate memory, walk through the inner linked list until you find
one item that is large enough (there should never be 2 available items
in a row so you don't need to check for that). Then, set it as used and
optionally split the end off into a seperate item.
To deallocate memory, walk through the inner linked list until you find
the item that points to the memory you want to deallocate. Set it to not
in use, and merge with the ones before and after if possible
(remembering to set header_in_use to false for any item that is no
longer in the chain.
To allocate a new item, walk through the outer linked list, and at each
stop iterate through every item at that stop, finishing when you find
one with header_in_use set to false. If you don't find any, you need to
allocate a new item in the outer linked list. Then set header_in_use to
true and use the inner item.
To deallocate an item, just set header_in_use to false. At the next
allocation, it will be found and used if needed.