Fill allocated memory with 0xCC. Also don't unmap NULL.
The first one is a security measure against re-reading another process's
freed memory. Additionally, it makes debugging multiple allocations simpler.

The second one is to let code unconditionally deallocate potentially-
uninitialized pointers. Without it, that would cause the *next*
allocation to fail somewhere higher up in the abstractions when it
checks for if malloc returned NULL.