Add begin() and end() wrapper functions
These wrap over arrays (with a template for the size) and objects that
have .data() and .size() member functions.

C++ has some magic with its for-loops that allow iterating over
containers that have .begin() and .end() member functions, but we don't
have any of those in the kernel, so I'm not worried about unifying those
as well.