Add full KVector class - with KString wrapper
Owning, editable, resizeable!
Aborts on out-of-bounds access.
Uses a c++ allocator, with KString using KGeneralAllocator<char>

Since it's a template, the implimentation is in the header file, except
for the check_index utility function that abort()s on out-of-bounds.
This is because it prints an error, and kernel/log.h will soon depend on
feline/kvector.h

Also add _kstr_vec for creating one from a string literal!