Don't call a virtual function in the terminal constructor
The chain was reset()->clear()->putchar(), which is virtual (shouldn't
compilers warn about this already). Because terminal (in particular
putchar) is pure (=0), the vtable entry for putchar points to
__cxa_pure_virtual (a function that shouldn't ever be called). I fixed
this by making it call move(0, 0) directly.

Because of this, I added a log+crash to __cxa_pure_virtual. According to
the comments I copied in from [https://wiki.osdev.org/C++]
(https://wiki.osdev.org/C%2BC%2B) it should never be called, but if gcc
and clang won't warn me about this, I need to add my own error checking.