Add rvalue overloads for non-member operator<<(kout&, ...) functions
It works find to do kLog() << type_with_member_operator<< because
reasons, and it returns a normal lvalue. Doing it for
type_with_nonmember_operator<< means the temporary fails to convert into
a regular lvalue, giving very weird errors. And since this isn't a
template, we can't use universal references! (The temporary is because
we *want* it to destruct at the end of the statement, and while making
it a const reference fixes the types, that only works with the NoAlloc()
versions.)

To make this simpler, I added a RVALUE_OVERLOAD macro which, given a
type, creates an operator<<(kout&&, type), which calls the lvalue
overload, and somehow the conversion is okay there!

TL;DR - after every operator<<(kout&, T) add RVALUE_OVERLOAD(T) and
things will just work.