Don't use exceptions in the kernel, but allow them when hosted
Adds the report_fatal_error(KStringView&) function which when hosted
throws an exception containing the error, allowing tests to catch and
still return true, and when in the kernel prints the error message
before calling abort(). The reason we don't just call abort either way
is because CTest has WILL_FAIL, but still doesn't like a test aborting,
so we need some way of testing errors we want to crash on.