Fix the ARM PMM (and actually unmap pages)
It was working from the incorrect assumption that every available memory
region would only be overlapped by only one unavailable memory region.
Not only was this incorrect, but by marking the kernel and the PMM as
using different locations, it was guaranteed to fail, leading to
printf being overwritten.

Now, the architecture-independant PMM does a bunch of work to mark
overlapping ranges as used, so lots of that code can be removed from
the ARM version.

While debugging this, it was discovered that the ARM unmap_page function
never actually did what its name described: it just marked the page as
unused in the searchable page table, but it was still mapped if you were
to write to it.