Avoid "this auto" to support GCC 13
It's a large diff from auto-formatting (why?), but
try_add_available_memory_region takes a reference to itself (must be
manually passed), instead of using `this auto` to make it happen
without the extra argument. The feature is technically supported in
c++23, but the version of GCC currently in the CI does not support it.