Add try_acquire_lock to the Spinlock class
This is useful for doing a task that needs a lock taken, but the task
might not need to be done if someone else already has the lock. Eg.
- task a calls sched()
- a timer interrupt happens, which generates a call to sched()
In this situation, using acquire_lock() would result in a deadlock,
but simply returning from sched() (and thus the interrupt handler)
means the existing call to sched() simply continues, which does what
we want.